Add support for Ubuntu
There are some slight differences between Debian and Ubuntu. Debian has /dev/shm already mounted, whereas Ubuntu, /dev/shm is a symbolic link to /run/shm - so removing it as a symbolic link and re-creating it.

Also, Ubuntu doesn't have the package chkconfig in the repository, which is used in the post install tasks of Oracle XE. So, we are creating a custom /sbin/chkconfig file based on installation steps described here: https://community.oracle.com/thread/2303639